+++ apps/pet-shop/src/app/integrator/data-access/backend-api/ps-categories/ps-categories-for-be.service.ts
@@ -3,9 +3,9 @@ import { inject, Injectable } from '@angular/core'
 import type { Observable } from 'rxjs'
 
 // # Internal modules
-import { environment } from '../../../../environments/environment'
+import { environment } from '../../../../../environments/environment'
 import { BackendAPIService } from '../backend-api/backend-api.service'
-import type { PSCategory } from './ps-categories-for-be.type'
+import type { PSCategoryForBE } from './ps-categories-for-be.type'
 
 @Injectable({
   providedIn: 'root',
@@ -13,7 +13,7 @@ import type { PSCategory } from './ps-categories-for-be.type'
 export class PSCategoriesForBEService {
   readonly #backendAPIService = inject(BackendAPIService)
 
-  public readList(): Observable<readonly PSCategory[]> {
-    return this.#backendAPIService.get<readonly PSCategory[]>(environment.API.categories.readList)
+  public readList(): Observable<readonly PSCategoryForBE[]> {
+    return this.#backendAPIService.get<readonly PSCategoryForBE[]>(environment.API.categories.readList)
   }
 }
