+++ apps/pet-shop/src/app/integrator/data-access/backend-api/ps-categories/ps-categories-for-be.type.ts
@@ -1,4 +1,4 @@
-export interface PSCategory {
+export interface PSCategoryForBE {
   readonly id: number
   readonly name: string
 }
